airflow-docker/requirements.txt
# Optimized requirements for Airflow Docker setup
# Only includes packages actually used in DAGs

# Core database connector
psycopg2-binary==2.9.9

# Data processing (used in data_processing_dag.py and file_operations_dag.py)
pandas==2.1.4

# Data generation library (used in file_operations_dag.py)
mimesis==15.1.0

# Airflow PostgreSQL provider (used in sql_basic_dag.py and data_processing_dag.py)
apache-airflow-providers-postgres==5.11.1

# Testing framework
pytest==7.4.4
